refactor: extract shared exercise list components
- Add useExerciseFilter hook for search/filter/grouping logic
- Add ExerciseFilterBar component (search + body part chips)
- Add ExerciseListView component (full SectionList with filter bar)
- ExerciseCard now shows body parts with primary bolded
- ExerciseList and ExerciseSelect reduced to thin wrappers
- EditRoutine uses shared hook + filter bar, supports duplicate exercises
- CreateRoutine updated for multi-body-part search
