+++ frontend/src/components/ExerciseCard.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { StyleSheet, View, TouchableOpacity } from "react-native";
 import { Text } from "@react-navigation/elements";
 import { useColorScheme } from "react-native";
+import { renderBodyParts } from "../utils/exerciseUtils";
 import { BodyPartDTO } from "../api/exerciseService";
 
 interface ExerciseCardProps {
@@ -27,6 +28,7 @@ export function ExerciseCard({
     subtle: isDark ? "#aaa" : "#666",
     border: isDark ? "#333" : "#e0e0e0",
     card: isDark ? "#1c1c1e" : "#fff",
+    accent: "#007AFF",
   };
 
   return (
@@ -49,6 +51,15 @@ export function ExerciseCard({
           >
             {exercise.name}
           </Text>
+          {exercise.bodyParts && exercise.bodyParts.length > 0 && (
+            <Text style={[styles.bodyPartsText, { color: colors.subtle }]}>
+              {renderBodyParts(
+                exercise.bodyParts,
+                colors.subtle,
+                colors.accent,
+              )}
+            </Text>
+          )}
           {exercise.description ? (
             <Text
               style={[styles.exerciseDescription, { color: colors.subtle }]}
@@ -87,6 +98,10 @@ const styles = StyleSheet.create({
   exerciseName: {
     fontSize: 16,
     fontWeight: "600",
+    marginBottom: 2,
+  },
+  bodyPartsText: {
+    fontSize: 12,
     marginBottom: 4,
   },
   exerciseDescription: {
